+++ libstmm-input-gtk/test/fakeflogtkbackend.h
@@ -79,7 +79,7 @@ public:
 	{
 		const int32_t nIdx = findXDeviceId(nXDeviceId);
 		assert(nIdx >= 0);
-		const int32_t nLastIdx = m_aXDeviceIds.size() - 1;
+		const int32_t nLastIdx = static_cast<int32_t>(m_aXDeviceIds.size()) - 1;
 		if (nIdx < nLastIdx) {
 			m_aXDeviceIds[nIdx] = m_aXDeviceIds[nLastIdx];
 		}
